+++ node_express/app/controllers/car.controller.js
@@ -1,8 +1,8 @@
-const { getCarsService } = require("../services/car.service");
+const { fetchCars } = require("../services/car.service");
 
 const getCars = async (req, res) => {
   try {
-    const cars = await getCarsService();
+    const cars = await fetchCars();
     res.json(cars);
   } catch (error) {
     res.status(500).send(error);