+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * [feat] Add node_express project.
 * [feat] Add gitignore in node_express project.
 * [feat] Add endpoint cars in node_express project.
+* [refactor] Service car in node_express project.
 
 ## v1.0.0